[fix] #53 checklist 중 범위 type 일치율 계산 로직 수정

+++ src/main/java/org/gachon/checkmate/domain/checkList/entity/HomeType.java
@@ -17,7 +17,7 @@ public enum HomeType implements EnumField {
     private final String desc;
     private final int size = 4;
 
-    public int compareRateTo(HomeType e) {
-        return 1 - Math.abs(Integer.parseInt(this.getCode()) - Integer.parseInt(e.getCode())) / (this.size - 1);
+    public double compareRateTo(HomeType e) {
+        return 1 - (double) Math.abs(Integer.parseInt(this.getCode()) - Integer.parseInt(e.getCode())) / (this.size - 1);
     }
 }
